+++ frontend/package.json
@@ -10,7 +10,6 @@
     "@testing-library/react": "^13.2.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^0.27.2",
-    "date-fns": "^2.30.0",
     "framer-motion": "6",
     "jwt-decode": "^3.1.2",
     "react": "^18.1.0",
