+++ frontend/src/index.js
@@ -5,8 +5,10 @@ import App from "./App";
 import "./index.css";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(< React.StrictMode >
-  <ChakraProvider >
-    <App />
-  </ChakraProvider> </React.StrictMode >
+root.render(
+  <React.StrictMode>
+    <ChakraProvider>
+      <App />
+    </ChakraProvider>
+  </React.StrictMode>
 );
