+++ frontend/src/services/axios.js
@@ -1,16 +1,16 @@
 import axios from "axios";
 
-const baseURL = "http://localhost:8000/api/v1/";
+const baseURL = "http://85.215.190.209:8000/api/v1/";
 
 const axiosInstance = axios.create({
-    baseURL,
+  baseURL,
 });
 
 axiosInstance.interceptors.response.use(
-    (response) => response,
-    (error) => {
-        return Promise.reject(error);
-    }
+  (response) => response,
+  (error) => {
+    return Promise.reject(error);
+  }
 );
 
 export default axiosInstance;
